Split replacer into replacement processor
