+++ snuba/query/project_extension.py
@@ -9,7 +9,7 @@ from snuba.query.expressions import Column, FunctionCall, Literal
 from snuba.query.extensions import QueryExtension
 from snuba.query.query import Query
 from snuba.query.query_processor import ExtensionData, ExtensionQueryProcessor
-from snuba.replacer import get_projects_query_flags
+from snuba.datasets.errors_replacer import get_projects_query_flags
 from snuba.request.request_settings import RequestSettings
 from snuba.state import get_config, get_configs
 from snuba.state.rate_limit import RateLimitParameters, PROJECT_RATE_LIMIT_NAME
